// server/src/Models/V1/Point/Interface/get.point.adjust.info.interface.ts
export interface GetPointAdjustInfoResp {
  adjustId: string;
  adjustName: string;
  consumeDate: string;
  adjustType: number;
  adjustPoint: number;
  dataType: number;
  adjustDate: string;
  memberType: number;
  point: number;
  mobileContryCode: string;
  mobile: string;
  memberId: string;
  name: string;
  fileUrl: string;
  fileName: string;
  fileCount: number;
  fileDataCount: number;
  activeStatus: number;
  activeDay: number;
  brandId: string;
  storeId: string;
  status: string;
  remark: string;
}